Simplify API, remove literal parsing
The locale parsing was somewhat tricky; For system timezones, we were naively
parsing a string like EST and using that. This is a problem if the system time
involved daylight savings -- e.g. EDT -- because our built-in locales did not
know about these. This caused a bug where trying to read these system times would
fail.

Fortunately, we were able to change the logic to avoid using the string and instead
use the system time directly. In doing so, it became clear that supporting literal tz
parsing like EST wasn't really worth it. We can just rely on tz database labels instead.

This has the added bonus of simplifying the API (only one way to pass in a tz) and code.

We also make other, minor changes:

- Remove rfc822 from format-in, as that includes tz info
- Fix readme examples
- Remove unused SrcTZ type
- Generate optics rather than hand-write them
- Add CallStack info to exceptions