Add windows to CI
Supporting windows revealed the fragility of the current
"system timezone parsing", which relied on reading a system timezone
like EST and adding that string to the string we ultimately want to
read. Windows formatted UTC as "Coordinated Universal Time", so we
would try to read e.g. "08:30 Coordinated Universal Time", which would
fail. Even when it worked, we had to manually add locale info
for the current system, which was flaky at best.

Thus we forget about TZ abbreviations and use Local's
'timeZoneOffsetString' function, which turns a timezone into a string
like "+1300". This requires no locale knowledge meaning we can remove
the flaky logic, thus increasing overall robustness.